[Editorial] brain fag -> brain fog
FadedPage uses "fog" here.  The original text had "fag," as in slang for fatigue.  However, "brain fog" is a modern phrase that has just about the same meaning as brain-fatigue, and it doesn't use the word "fag" which in modern speech, most commonly is a slur for a homosexual person. I would prefer not to treat this as a typo, but rather an editorial choice.