[Editorial] Nine children, so tenth -> ninth
If you read the paragraph where this change occurs, you will see that
the mother has nine children.  It makes more sense if Anne takes pity
on the mother and allows the last (ninth) child to live.

This change is often made on newer books. A search of books hosted on
the Internet Archive shows ten different versions of the book that
include this change (the search shows 12 hits, but the oldest three
are duplicates). The Internet Archive has 16 copies of Anne of the
Island that are English language texts published after 1980. Nine of
the sixteen books have this change.