[Editorial] Fix typo, argured -> augured
In the page scans, the book has "argured," which is not in the
dictionary. Although both "argued" and "augured" are reasonable
corrections, I think "augured" is a better fit because it has the
sense of foretelling, or predicting the future. Also, according to
Google Ngram, "augured well" shows up about five times as often as
the phrase "argued well," although that doesn't tell us anything
about the context. This change is frequently made on newer books.